Set canvas size convenient for PDF / SVG export.
When exporting into pdf of svg the canvas in the export will be set to
something like 744x1052 (probably taken from svg tag attributes in file
`app/pencil-core/exporter/Pencil2SVG.xslt` line 43, cf.
https://github.com/evolus/pencil/issues/213).

This change adds an option in the canvas (resize) menu to resize the canvas
to this size.