Fix Electron's Windows Configuration
The "win" key for electron configuration in the package.json file
contains outdated keys, which results in an error when building a
release with electron-builder.

These changes fix the issue by moving the description & author fields
into top-level keys, changing the iconUrl key to icon, and removing
the unsupported owners & copyright fields.